Separate windows and unix code
Move unix specific code to unix only build files. Add
stub implementation for Windows where needed. Remove
syscall importing from platform-agnostic files.
