+++ driver.go
@@ -1,7 +1,6 @@
 package continuity
 
 import (
-	"errors"
 	"os"
 	"strconv"
 )
@@ -142,17 +141,4 @@ func (d *driver) Lchown(name, uidStr, gidStr string) error {
 
 func (d *driver) Symlink(oldname, newname string) error {
 	return os.Symlink(oldname, newname)
-}
-
-func (d *driver) Mknod(path string, mode os.FileMode, major, minor int) error {
-	return mknod(path, mode, major, minor)
-}
-
-func (d *driver) Mkfifo(path string, mode os.FileMode) error {
-	if mode&os.ModeNamedPipe == 0 {
-		return errors.New("mode passed to Mkfifo does not have the named pipe bit set")
-	}
-	// mknod with a mode that has ModeNamedPipe set creates a fifo, not a
-	// device.
-	return mknod(path, mode, 0, 0)
 }
