+++ hardlinks_windows.go
@@ -1,5 +1,12 @@
 package continuity
 
-// NOTE(stevvooe): Obviously, this is not yet implemented. However, the
-// makings of an implementation are available in src/os/types_windows.go. More
-// investigation needs to be done to figure out exactly how to do this.
+import "os"
+
+type hardlinkKey struct{}
+
+func newHardlinkKey(fi os.FileInfo) (hardlinkKey, error) {
+	// NOTE(stevvooe): Obviously, this is not yet implemented. However, the
+	// makings of an implementation are available in src/os/types_windows.go. More
+	// investigation needs to be done to figure out exactly how to do this.
+	return hardlinkKey{}, errNotAHardLink
+}
